refactor: Extract WandAction and use it to refine prompts
